requirements.txt
# Moviqo Backend Dependencies

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.10.0
joblib>=1.3.0

# Production Server
gunicorn>=21.0.0
